Time: 56 ms (10.10%) | Memory: 21.4 MB (28.28%) - LeetSync

+++ 907-koko-eating-bananas/koko-eating-bananas.cpp
@@ -17,7 +17,7 @@ public:
             hr=h;
             cout<< mid << " " << k << " "<<start<< " "<<end << endl;
             for(auto i :piles){
-                hr=hr-ceil((double)i/(double)mid);
+                hr=hr-ceil(i/(double)mid);
                 if(hr<0)break;
             }   
             if(hr>=0){
